# docs/interfaces/openapi_v1.yaml
...
openapi: 3.0.3
info:
  title: PitWall API
  version: 0.1.0
  description: |
    Real-time F1 undercut detection engine.

    The HTTP surface exposes the data and control plane:
    - **Catalogue**: list available sessions, fetch fitted degradation curves.
    - **Replay control**: start and stop a deterministic replay of a
      historical session through the in-memory engine.
    - **Snapshots**: read the engine's race state.
    - **Configuration**: switch the active pace predictor at runtime.
    - **Back-test**: compare alerts emitted by the engine against the
      curated list of known historical undercuts.

    The push surface (`/ws/v1/live`) is a WebSocket channel; see
    `docs/interfaces/websocket_messages.md` for the message envelope and
    the supported `type` values.

    See also `docs/architecture.md`, ADR 0008 (OpenAPI as source of truth),
    and quanta 04 (window-of-undercut formula).

  contact:
    name: PitWall Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local dev (host-side)
  - url: http://backend:8000
    description: Inside the docker-compose network
...
paths:
...
  /api/v1/config/predictor:
    post:
      tags: [config]
      operationId: setActivePredictor
      summary: Switch the active pace predictor at runtime
      description: |
        Switches `PacePredictor` between `scipy` and `xgboost` without
        restarting the backend. The new predictor takes effect on the
        next `lap_complete` processed by the engine. 409 if the
        requested predictor's model is not loaded (e.g. `xgboost`
        before `make train-xgb` has run).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetPredictorRequest" }
            examples:
              switch_to_xgb:
                value: { predictor: xgboost }
      responses:
        "200":
          description: OK. Returns the now-active predictor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetPredictorResponse" }
        "400":
          description: Unknown predictor name.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: Requested predictor is not loaded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
components:
...
  schemas:
...
    PredictorName:
      type: string
      enum: [scipy, xgboost]
...
    Problem:
      type: object
      required: [detail]
      description: |
        FastAPI's default error envelope. Kept minimal in V1; can be
        upgraded to RFC 7807 (`type`, `title`, `status`) without
        breaking clients because `detail` is preserved.
      properties:
        detail:
          oneOf:
            - { type: string }
            - { type: array, items: { type: object } }
          example: Session 'monaco_2024_R' not found.
...
    SetPredictorRequest:
      type: object
      required: [predictor]
      properties:
        predictor: { $ref: "#/components/schemas/PredictorName" }
    SetPredictorResponse:
      type: object
      required: [active_predictor]
      properties:
        active_predictor: { $ref: "#/components/schemas/PredictorName" }
...